+++ Assets/IngredientShopOffer.cs
@@ -5,7 +5,7 @@ public class IngredientShopOffer : ShopOffer
 {
     [SerializeField] private PotionIngredient label;
 
-    public void Set(PotionIngredient setTo, float cost, Action onClick)
+    public void Set(PotionIngredient setTo, int cost, Action onClick)
     {
         label = setTo;
         this.cost = cost;
